+++ rust/candid/src/types/internal.rs
@@ -374,7 +374,7 @@ impl std::fmt::Display for Label {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Label::Id(n) | Label::Unnamed(n) => {
-                write!(f, "{}", super::number::pp_num_str(&n.to_string()))
+                write!(f, "{}", crate::pretty_printer::pp_num_str(&n.to_string()))
             }
             Label::Named(id) => write!(f, "{id}"),
         }